Revert "LPS-168476 Extract component"
This reverts commit b6d3ac01341aea708139f7b9bf7f5f2bf4b4404b.
